+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod backend;
 pub mod cloud_hypervisor;
 pub mod config;
 pub mod console;